Remove context set/get from the API
