+++ lib/surface/compiler/eex_engine.ex
@@ -132,7 +132,7 @@ defmodule Surface.Compiler.EExEngine do
 
     props_expr =
       quote generated: true do
-        Keyword.merge(unquote(props_expr), __context__: @__context__, __context2__: @__context2__)
+        Keyword.merge(unquote(props_expr), __context__: @__context__)
       end
 
     # TODO: map names somehow?
@@ -213,8 +213,7 @@ defmodule Surface.Compiler.EExEngine do
           @socket,
           unquote(module),
           Surface.build_assigns(
-            @__context__ || [],
-            @__context2__ || %{},
+            @__context__ || %{},
             unquote(props_expr),
             unquote(dynamic_props_expr),
             unquote(slot_props),
@@ -230,8 +229,7 @@ defmodule Surface.Compiler.EExEngine do
           @socket,
           unquote(module),
           Surface.build_assigns(
-            @__context__ || [],
-            @__context2__ || %{},
+            @__context__ || %{},
             unquote(props_expr),
             unquote(dynamic_props_expr),
             unquote(slot_props),
