+++ lib/surface/content_handler.ex
@@ -74,7 +74,7 @@ defmodule Surface.ContentHandler do
 
       # TODO [Context]: Update this to be only the appropriate assigns for context
       case args[:__context__] || assigns.__context__ do
-        [] ->
+        %{} ->
           assigns.inner_content.(Keyword.merge(prop_assigns, __slot__: {name, index}))
 
         context_assign ->
@@ -89,7 +89,6 @@ defmodule Surface.ContentHandler do
     fn args ->
       # TODO [Context]: Update this to be only the appropriate assigns for context
       context_assign = args[:__context__] || assigns.__context__
-      context2_assign = args[:__context2__] || assigns.__context2__
 
       prop_assigns =
         Enum.map(all_prop_assign_mappings, fn mappings_for_index ->
@@ -98,12 +97,11 @@ defmodule Surface.ContentHandler do
           end)
         end)
 
-      join_contents(assigns, size, context_assign, context2_assign, prop_assigns)
+      join_contents(assigns, size, context_assign, prop_assigns)
     end
   end
 
-  defp join_contents(assigns, 1, [], context2_assign, assign_mappings)
-       when context2_assign == %{} do
+  defp join_contents(assigns, 1, context_assign, assign_mappings) when context_assign == %{} do
     if assigns[:inner_content] == nil do
       ""
     else
@@ -115,10 +113,10 @@ defmodule Surface.ContentHandler do
     end
   end
 
-  defp join_contents(assigns, size, context_assign, context2_assign, assign_mappings) do
+  defp join_contents(assigns, size, context_assign, assign_mappings) do
     ~L"""
     <%= if assigns[:inner_content] != nil do %>
-    <%= for index <- 0..size-1 do %><%= assigns.inner_content.(Keyword.merge(Enum.at(assign_mappings, index), [__slot__: {:__default__, index}, __context__: context_assign, __context2__: context2_assign])) %><% end %>
+    <%= for index <- 0..size-1 do %><%= assigns.inner_content.(Keyword.merge(Enum.at(assign_mappings, index), [__slot__: {:__default__, index}, __context__: context_assign])) %><% end %>
     <% end %>
     """
   end
