+++ test/api_test.exs
@@ -1,24 +1,6 @@
 defmodule Surface.APITest do
   use ExUnit.Case, async: true
 
-  defmodule ContextSetter do
-    use Surface.Component
-
-    @doc """
-    The Form struct defined by the parent <Form/> component.
-    """
-    context set form, :form
-
-    def init_context(_assigns) do
-      {:ok, form: :fake}
-    end
-
-    def render(assigns) do
-      ~H"""
-      """
-    end
-  end
-
   test "raise error at the right line" do
     code = "property label, :unknown_type"
     message = ~r/code:4/
@@ -91,56 +73,11 @@ defmodule Surface.APITest do
 
     code = """
     data label, :string
-    context get label, from: Surface.APITest.ContextSetter
+    property label, :string
     """
 
-    message = ~r"""
-    cannot use name "label". There's already a data assign with the same name at line 4.
-
-    Hint: you can use the :as option to set another name for the context assign.\
-    """
-
+    message = ~r/cannot use name "label". There's already a data assign/
     assert_raise(CompileError, message, fn -> eval(code) end)
-
-    code = """
-    property form, :form
-    context set form, :form
-    """
-
-    message = ~r"""
-    cannot use name "form". There's already a property assign with the same name at line 4.
-
-    Hint: if you only need this context assign in the child components, \
-    you can set option :scope as :only_children to solve the issue.\
-    """
-
-    assert_raise(CompileError, message, fn -> eval(code) end)
-
-    code = """
-    context set form, :form
-    property form, :form
-    """
-
-    message = ~r/cannot use name "form". There's already a context/
-    assert_raise(CompileError, message, fn -> eval(code) end)
-  end
-
-  test "allow context :set with existing assign name when :scope is :only_children" do
-    code = """
-    property form, :form
-    context set form, :form, scope: :only_children
-    """
-
-    assert {:ok, _} = eval(code)
-  end
-
-  test "allow context :get with existing assign name when using :as" do
-    code = """
-    property form, :form
-    context get form, from: Surface.APITest.ContextSetter, as: :my_form
-    """
-
-    assert {:ok, _} = eval(code)
   end
 
   test "accept invalid quoted expressions like literal maps as default value" do
@@ -308,211 +245,6 @@ defmodule Surface.APITest do
     end
   end
 
-  describe "context :set" do
-    test "validate action without options" do
-      code = "context unknown name, :string"
-      message = ~r/invalid context action. Expected :get or :set, got: :unknown/
-
-      assert_raise(CompileError, message, fn ->
-        eval(code)
-      end)
-    end
-
-    test "validate action with options" do
-      code = "context unknown name, :string, scope: :only_children"
-
-      message = ~r"""
-      invalid use of context. Usage: `context get name, opts` or \
-      `context set name, type, opts \\ \[\]`\
-      """
-
-      assert_raise(CompileError, message, fn ->
-        eval(code)
-      end)
-    end
-
-    test "validate name" do
-      code = "context set {a, b}, :string"
-      message = ~r/invalid context name. Expected a variable name, got: {a, b}/
-
-      assert_raise(CompileError, message, fn ->
-        eval(code)
-      end)
-    end
-
-    test "validate type" do
-      code = "context set name, scope: :only_children"
-      message = ~r/no type defined for context set. Type is required after the name./
-
-      assert_raise(CompileError, message, fn ->
-        eval(code)
-      end)
-
-      code = "context set name"
-
-      assert_raise(CompileError, message, fn ->
-        eval(code)
-      end)
-    end
-
-    test "valid options" do
-      code = "context set field, :atom, scope: :only_children"
-      assert {:ok, _} = eval(code)
-    end
-
-    test "validate :scope" do
-      code = "context set field, :atom, scope: :unknown"
-
-      message =
-        ~r/invalid value for option :scope. Expected :only_children or :self_and_children, got: :unknown/
-
-      assert_raise(CompileError, message, fn ->
-        eval(code)
-      end)
-    end
-
-    test "no required options" do
-      code = "context set field, :atom"
-      assert {:ok, _} = eval(code)
-    end
-
-    test "unknown options" do
-      code = "context set label, :string, a: 1"
-      message = ~r/unknown option :a. Available options: \[:scope\]/
-
-      assert_raise(CompileError, message, fn ->
-        eval(code)
-      end)
-    end
-  end
-
-  describe "context :get" do
-    test "validate action" do
-      code = "context unknown name, from: Surface.APITest.ContextSetter"
-      message = ~r/invalid context action. Expected :get or :set, got: :unknown/
-
-      assert_raise(CompileError, message, fn ->
-        eval(code)
-      end)
-    end
-
-    test "valid options" do
-      code = """
-      context get form, from: Surface.APITest.ContextSetter, as: :my_form
-      """
-
-      assert {:ok, _} = eval(code)
-    end
-
-    test "invalid options" do
-      code = """
-      context get form, {:a, :b}
-      """
-
-      message = ~r"""
-      invalid options for context form. Expected a keyword list of options, got: {:a, :b}\
-      """
-
-      assert_raise(CompileError, message, fn ->
-        eval(code)
-      end)
-
-      code = """
-      context get form, [:a, :b]
-      """
-
-      message = ~r"""
-      invalid options for context form. Expected a keyword list of options, got: \[:a, :b\]\
-      """
-
-      assert_raise(CompileError, message, fn ->
-        eval(code)
-      end)
-    end
-
-    test "no name defined" do
-      code = """
-      context get
-      """
-
-      message = ~r"""
-      no name defined for context get
-
-      Usage: context get name, opts
-      """
-
-      assert_raise(CompileError, message, fn ->
-        eval(code)
-      end)
-    end
-
-    test "invalid :from" do
-      code = """
-      context get form, from: 1
-      """
-
-      message = ~r/invalid value for option :from. Expected a module, got: 1/
-
-      assert_raise(CompileError, message, fn ->
-        eval(code)
-      end)
-    end
-
-    test "option :from is required" do
-      code = """
-      context get form, as: :my_form
-      """
-
-      message = ~r/the following options are required: \[:from\]/
-
-      assert_raise(CompileError, message, fn ->
-        eval(code)
-      end)
-
-      code = """
-      context get form
-      """
-
-      assert_raise(CompileError, message, fn ->
-        eval(code)
-      end)
-    end
-
-    test "invalid :as" do
-      code = """
-      context get form, from: Surface.APITest.ContextSetter, as: 1
-      """
-
-      message = ~r/invalid value for option :as. Expected an atom, got: 1/
-
-      assert_raise(CompileError, message, fn ->
-        eval(code)
-      end)
-    end
-
-    test "cannot define the type of the assign" do
-      code = "context get name, :string"
-
-      message = ~r"""
-      cannot redefine the type of the assign when using action :get. \
-      The type is already defined by a parent component using action :set\
-      """
-
-      assert_raise(CompileError, message, fn ->
-        eval(code)
-      end)
-    end
-
-    test "unknown options" do
-      code = "context get label, from: Surface.APITest.ContextSetter, a: 1"
-      message = ~r/unknown option :a. Available options: \[:from, :as\]/
-
-      assert_raise(CompileError, message, fn ->
-        eval(code)
-      end)
-    end
-  end
-
   test "generate documentation when no @moduledoc is defined" do
     assert get_docs(Surface.PropertiesTest.Components.MyComponent) == """
            ### Properties
@@ -546,10 +278,6 @@ defmodule Surface.APITest do
       use Surface.LiveComponent
 
       #{code}
-
-      def init_context(_assigns) do
-        {:ok, field: nil}
-      end
 
       def render(assigns) do
         ~H(<div></div>)
@@ -760,37 +488,6 @@ defmodule Surface.APISyncTest do
       assert output =~ ~r"""
              cannot render <NonExisting> \(module NonExisting could not be loaded\)
                code.exs:7:\
-             """
-    end
-  end
-
-  describe "context :set error/warnings" do
-    test "warn on context :set if there's no init_context/1" do
-      id = :erlang.unique_integer([:positive]) |> to_string()
-      module = "TestLiveComponent_#{id}"
-
-      code = """
-      defmodule #{module} do
-        use Surface.LiveComponent
-
-        context set field, :atom
-
-        def render(assigns) do
-          ~H(<div></div>)
-        end
-      end
-      """
-
-      output =
-        capture_io(:standard_error, fn ->
-          {{:module, _, _, _}, _} =
-            Code.eval_string(code, [], %{__ENV__ | file: "code.exs", line: 1})
-        end)
-
-      assert output =~ ~r"""
-             context assign "field" not initialized. You should implement an init_context/1 \
-             callback and initialize its value by returning {:ok, field: ...}
-               code.exs:4:\
              """
     end
   end
