+++ test/context_through_slot_test.exs
@@ -8,17 +8,13 @@ defmodule Surface.ContextThroughSlotTest do
     use Surface.Component
 
     property foo, :string
-    context set foo, :string, scope: :only_children
     slot default
-
-    # The foo prop is passed here and so we can use it
-    def init_context(assigns) do
-      {:ok, foo: assigns.foo}
-    end
 
     def render(assigns) do
       ~H"""
-      <slot />
+      <Context set={{ :foo, @foo }}>
+        <slot />
+      </Context>
       """
     end
   end
@@ -42,12 +38,12 @@ defmodule Surface.ContextThroughSlotTest do
   defmodule Child do
     use Surface.Component
 
-    context get foo, from: Parent.ContextProvider
-
     def render(assigns) do
       # @foo is nil here
       ~H"""
-      <div>{{ @foo  }}</div>
+      <Context get={{ :foo }}>
+        <div>{{ @foo }}</div>
+      </Context>
       """
     end
   end
